Nvme subsystem minimal support
To support multipath capable drives we need to get their physical location.
This method should be sufficient for the drives connected via one port.

Supporting dual-ports and full multipathing support (especially in
ledmon) is a problem that requires better implementation. This is low risk
effort to make it recognizable in ledctl and in library in standard cases.

Instead parsing and reparsing, try to determine devnode on
block_device_init and save it. We can refer to this property in
led_device_name_lookup(). It is fair to require context for lookup now.

It removes possibility to use sysfs path in set-slot by device - it is
fair requirement.

Signed-off-by: Mariusz Tkaczyk <mariusz.tkaczyk@linux.intel.com>